added lookup cache for cache annotations
Because now the CacheFilter has too loop over every annotation on an IpcCommand
(which, depending on how many other annotations it has, may last very long),
the real cache annotation on the command is saved in a ConcurrentMap.
Also changed annotation.getClass() into annotation.annotationType() (thanks to Willi for that tip).